requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cryptography==2.6.1
entrypoints==0.3
Flask==1.1.1
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
paho-mqtt==1.5.0
picamera==1.13
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4
pyxdg==0.25
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.7
urllib3==1.24.1
Werkzeug==0.16.0
